saveData: re-organise
