+++ src/comp/0-YearPlanner/useYearPlanner.ts
@@ -1,0 +1,5 @@
+const useYearPlanner = function () {
+  return { };
+};
+
+export default useYearPlanner;
