+++ src/comp/2-body/calendar/comps.tsx
@@ -1,6 +1,6 @@
-import Cell from "../../lib/cell/Cell";
-import { MONTH_SHORT_NAME } from "../../util/constant";
-import { MONTH_LABEL_MIN_WIDTH } from "./Planner";
+import Cell from "../../../lib/cell/Cell";
+import { MONTH_SHORT_NAME } from "../../../util/constant";
+import { MONTH_LABEL_MIN_WIDTH } from "./Calendar";
 import MonthLabel from "../month-label/MonthLabel";
 import { StyledDayVoid } from "../day/Day.style";
