+++ src/comp/day-modal/DayModal.tsx
@@ -4,9 +4,10 @@ import { BOOLEAN_STR_VALUES } from "../../data-lib/util/constant";
 import TextField from "../../data-lib/adapter/MU-adapter/textField/TextField";
 import { useContext } from "react";
 import { DataContext } from "../../data-lib/context/DataProvider";
-import useYearPlanner from "../1-YearPlanner/useYearPlanner";
+import useYearPlanner from "../0-YearPlanner/useYearPlanner";
 import { paths } from "../../util/constant";
 import useComp, { IComp } from "../../data-lib/hook/useComp";
+import useCommon from '../../hook/useCommon';
 
 const TxtDayNote = () => {
   const props: IComp = {
@@ -45,7 +46,7 @@ const TxtDayNote = () => {
 function DayModal(props: any) {
   const { getValue, setValue } = useContext(DataContext);
   const modalData = getValue ? getValue(paths.temp.dayModal._path) : {};
-  const { saveDate } = useYearPlanner();
+  const { saveDate } = useCommon();
 
   const onConfirm = () => {
     saveData();
